fix error in alembic.ini

+++ models/user.py
@@ -11,7 +11,7 @@ user = sqlalchemy.Table(
     sqlalchemy.Column("password", sqlalchemy.String(255)),
     sqlalchemy.Column("first_name", sqlalchemy.String(200)),
     sqlalchemy.Column("last_name", sqlalchemy.String(200)),
-    sqlalchemy.Column("phone", sqlalchemy.Integer(20)),
+    sqlalchemy.Column("phone", sqlalchemy.Integer),
     sqlalchemy.Column("role", sqlalchemy.Enum(RoleType), nullable=False,
                       server_default=RoleType.complainer.name),
     sqlalchemy.Column("iban", sqlalchemy.String(200)),
